XWIKI-966 New fix for Lucene delete both for performance and consistency. Also fixed the lucene debug logging
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/branches/XWIKI_1_0@2850 f329d543-caf0-0310-9063-dda96c69346f

+++ core/src/main/java/com/xpn/xwiki/plugin/lucene/IndexUpdater.java
@@ -238,7 +238,7 @@ public class IndexUpdater implements Runnable, XWikiDocChangeNotificationInterfa
                 LOG.debug("delete doc " + id);
             }
             try {
-                reader.deleteDocument(id);
+                reader.deleteDocument(id.intValue());
                 nb++;
             } catch (IOException e1) {
                 LOG.error("error deleting doc " + id, e1);
